+++ chats/apps/queues/models.py
@@ -4,7 +4,9 @@ from django.core.exceptions import ObjectDoesNotExist
 from django.db import models
 from django.utils import timezone
 from django.utils.translation import gettext_lazy as _
+from django.db.models import Subquery, OuterRef, Q
 
+from chats.apps.projects.models.models import CustomStatus
 from chats.core.models import BaseConfigurableModel, BaseModel, BaseSoftDeleteModel
 
 from .queue_managers import QueueManager
@@ -71,13 +73,17 @@ class Queue(BaseSoftDeleteModel, BaseConfigurableModel, BaseModel):
             project_permissions__queue_authorizations__role=1,
         )
 
-        # remove agents that have active custom status other than in service
+        custom_status_query = Subquery(
+            CustomStatus.objects.filter(
+                Q(user__id=OuterRef("id"))
+                & Q(is_active=True)
+                & Q(status_type__project=self.sector.project)
+                & ~Q(status_type__name__iexact="in-service")
+            ).values("user__id")
+        )
+
         return agents.exclude(
-            models.Q(
-                user_custom_status__is_active=True,
-                user_custom_status__project=self.sector.project,
-            )
-            & ~models.Q(user_custom_status__status_type__name__iexact="In-service")
+            id__in=custom_status_query
         )  # TODO: Set this variable to ProjectPermission.STATUS_ONLINE
 
     @property